Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/image-redirect-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8000/security/resources/abe.png because it does not appear in the img-src directive of the Content Security Policy.
Tests that a cross-origin image loaded via a redirect is blocked by the Content Security Policy. This test PASSED if there is a console warning message.
